Updated to only have fixed algorithm, next task is clean up
