+++ src/fit_pieces.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   fit_pieces.c                                       :+:      :+:    :+:   */
+/*   new_fit_pieces.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kdavis <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2016/10/12 10:29:43 by kdavis            #+#    #+#             */
-/*   Updated: 2016/10/21 11:14:58 by kdavis           ###   ########.fr       */
+/*   Created: 2016/10/22 11:02:34 by kdavis            #+#    #+#             */
+/*   Updated: 2016/10/24 14:44:21 by kdavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,134 +14,121 @@
 #include "fillit.h"
 
 /*
+** Checks to see if the piece values are at the right border of the puzzle.
+** if they are, move pieces back to far left as well as increment map
+** to the next row.
+*/
+
+t_puzz	*move_piece(t_piece *p, t_puzz *l)
+{
+	int i;
+
+	i = -1;
+	if (l->shift >= l->sq_size)
+	{
+		while (++i < 4)
+			p[l->pu].r[i] = p[l->pu].a[i];
+		l->shift = 0;
+		l->mrow++;
+	}
+	else
+	{
+		while (++i < 4)
+			p[l->pu].r[i] = p[l->pu].r[i] << 1;
+			l->shift++;
+	}
+	p[l->pu].mrow = l->mrow;
+	p[l->pu].shift = l->shift;
+	return (fit_piece(p, l));
+}
+
+/*
 ** Edit_map will either cut or paste a piece depending on the status.
 ** status = 1 indicates that the piece was  pasted into the map.
 ** status = 0 indicates that the piece was cut from the map.
 */
 
-t_piece		edit_map(t_piece p, t_ull *map, t_puzz l, int status)
+t_puzz	*edit_map(t_piece *p, t_puzz *l, int status)
 {
 	int	i;
 
 	i = -1;
+	if (status == 0)
+	{
+		l->pu--;
+		l->mrow = p[l->pu].mrow;
+	}
 	while (++i < 4)
-	{
-		map[l.mrow + i] = p.r[i] ^ map[l.mrow + i];
-	}
-	p.placed = status;
-	if (p.placed)
-	{
-		l.po[0]++;
-		p.order = *l.po;
-	}
-	else
-	{
-		l.po[0]--;
-		p.order = 0;
-	}
-	return (p);
+		l->map[l->mrow + i] = p[l->pu].r[i] ^ l->map[l->mrow + i];
+	if (status == 0)
+		return (l);
+	l->pu++;
+	return (fit_piece(p, l));
 }
 
 /*
-** Reset pieces will move all unused pieces to the horizontal location that
-** the map was reset from. It will also set the g_shift to reflect the new
-** horizontal position of all pieces.
+** Starting_position determines where in the map to start trying to place 
+** the next piece. Normally the starting position is at the top left of the
+** map unless the piece is a duplicate of an already placed piece. In that
+** case the piece starts after the already placed piece.
 */
 
-t_piece		*reset_pieces(t_piece *p, t_puzz l)
+t_puzz	*starting_position(t_piece *p, t_puzz *l)
 {
-	int	pu;
 	int	i;
+	int	pri;
+	int	counter;
 
-	pu = -1;
-	while (++pu < l.pnbr)
+	i = -1;
+	if (p[l->pu].placed == 0)
 	{
-		if (p[pu].placed == 0)
+		l->shift = 0;
+		l->mrow = 0;
+	}
+	while (++i < l->pnbr)
+	{
+		pri = -1;
+		counter = 0;
+		if (p[i].placed == 1)
+			while (++pri < 4)
+				if (p[l->pu].a[pri] == p[i].a[pri])
+					counter++;
+		if (counter == 4)
 		{
-			i = -1;
-			while (++i < 4)
-				p[pu].r[i] = p[pu].a[i] << l.shift;
+			l->shift = p[i].shift;
+			l->mrow = p[i].mrow;
 		}
 	}
-	return (p);
-}
-
-/*
-** Adjust_coord adjusts the paste coordinates for pieces such as the S
-** piece that will paste earlier than designated empty space should be.
-*/
-
-t_puzz		adjust_coord(t_piece p, t_puzz l, t_ull *map)
-{
-	int	row;
-	int	cs;
-
-	row = p.a[0];
-	cs = 1 << l.shift;
-	row = row ^ 15;
-	if (cs & map[l.mrow])
-		while ((row & 1))
-		{
-			l.shift++;
-			row = row >> 1;
-		}
+	p[l->pu].shift = l->shift;
+	p[l->pu].mrow = l->mrow;
 	return (l);
 }
 
 /*
-** Pieces_left calculates how many pieces have still not been placed
-** into the map yet.
+** Fit_piece works recursivley by finding the perfect position for each piece
+** in order. If a piece is pasted we move on to the next piece, if we can not
+** paste the piece then we move to the next position.
 */
 
-int			pieces_left(t_piece *p, t_puzz l)
+t_puzz	*fit_piece(t_piece *p, t_puzz *l)
 {
-	int	pu;
-	int	pl;
+	t_puzz *solved;
 
-	pu = 0;
-	pl = 0;
-	while (pu < l.pnbr)
-		if (p[pu++].placed == 0)
-			pl++;
-	return (pl);
-}
-
-/*
-** Main body of the recursive function.
-** 117. Calculates the number of pieces left.
-** 118-119: Selects the first unused piece to fit into the puzzle.
-** 122-138: Trying to fit pieces into puzzle as long as there are pieces left
-**			and the maximum number of empty blocks has not been exceeded.
-**	If the piece fits then add it to the map and move to the next map position.
-**	If the piece was placed in the incorrect position remove it from the map
-**		and reset all unused pieces to the appropriate horizontal value.
-*/
-
-t_ull		*fit_pieces(t_piece *p, t_ull *map, int eb_nbr, t_puzz legend)
-{
-	t_ull	*tail;
-	int		pu;
-	int		pl;
-
-	pl = pieces_left(p, legend);
-	if ((pu = starting_piece(p, legend)) == 42)
-		return (map);
-	while (++pu < legend.pnbr && eb_nbr <= legend.eb_max)
-		if ((p[pu].placed) != 1)
-		{
-			if ((check_map(p[pu], map, legend)))
-			{
-				p[pu] = edit_map(p[pu], map, legend, 1);
-				legend = adjust_coord(p[pu], legend, map);
-			}
-			if (p[pu].placed == 1 || --pl == 0)
-				if ((tail = move_bits(p, map, eb_nbr, legend)))
-					return (tail);
-			if (p[pu].placed == 1)
-			{
-				p[pu] = edit_map(p[pu], map, legend, 0);
-				p = reset_pieces(p, legend);
-			}
-		}
+	if (l->pu == l->pnbr)
+		return (l);
+	l = starting_position(p, l);
+	p[l->pu].placed = 1;
+	if (check_map(p[l->pu], l->map, *l))
+	{
+		if ((solved = edit_map(p, l, 1)))
+			return (solved);
+		l = edit_map(p, l, 0);
+	}
+	if ((check_bottom(p[l->pu], *l)) || (check_border(p[l->pu], *l)))
+	{
+		if ((solved = move_piece(p, l)))
+			return (solved);
+		p[l->pu].placed = 0;
+	}
 	return (NULL);
 }
