+++ src/map_check.c
@@ -6,7 +6,7 @@
 /*   By: kdavis <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/17 10:26:52 by kdavis            #+#    #+#             */
-/*   Updated: 2016/10/22 21:20:00 by kdavis           ###   ########.fr       */
+/*   Updated: 2016/10/24 14:50:05 by kdavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
