+++ src/map_maker.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   map_maker.c                                        :+:      :+:    :+:   */
+/*   new_map_maker.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kdavis <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2016/10/18 09:54:14 by kdavis            #+#    #+#             */
-/*   Updated: 2016/10/22 14:11:48 by kdavis           ###   ########.fr       */
+/*   Created: 2016/10/22 20:20:42 by kdavis            #+#    #+#             */
+/*   Updated: 2016/10/24 14:52:41 by kdavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,29 +57,26 @@ t_piece	*total_reset(t_piece *pieces, t_puzz l)
 ** pieces can't fit in the previous square.
 */
 
-t_puzz	eb_calculator(t_piece *pieces, t_puzz l)
+t_puzz	*eb_calculator(t_piece *pieces, t_puzz *l)
 {
-	t_ull	*solved;
-//	int		eb_nbr;
+	t_puzz	*solved;
 	int		pu;
 
 	pu = 0;
-//	eb_nbr = 0;
-	l.eb_max = (l.sq_size * l.sq_size) - (4 * l.pnbr);
-	l.mrow = 0;
-	l.shift = 0;
-	l.pu = 0;
-	if (!(l.map = (t_ull *)ft_memalloc(sizeof(t_ull) * (l.sq_size + 3))))
-		exit(-1);
-	if ((solved = fit_piece/*s*/(pieces, /*l.map, eb_nbr,*/ l)))
+	l->mrow = 0;
+	l->shift = 0;
+	l->pu = 0;
+	if (!(l->map = (t_ull *)ft_memalloc(sizeof(t_ull) * (l->sq_size + 3))))
+		return (NULL);
+	if ((solved = fit_piece(pieces, l)))
 	{
-		pieces = sort_pieces(pieces, l);
-		print_solution(pieces, solved, l);
+		pieces = sort_pieces(pieces, *solved);
+		print_solution(pieces, solved->map, *solved);
 		return (l);
 	}
-	pieces = total_reset(pieces, l);
-	ft_memdel((void *)&l.map);
-	l.sq_size++;
+	pieces = total_reset(pieces, *l);
+	ft_memdel((void *)&l->map);
+	l->sq_size++;
 	l = eb_calculator(pieces, l);
 	return (l);
 }
@@ -88,9 +85,9 @@ t_puzz	eb_calculator(t_piece *pieces, t_puzz l)
 ** Initializes the square size and activates the recursion loop.
 */
 
-void	map_maker(t_piece *pieces, t_puzz legend)
+void	map_maker(t_piece *pieces, t_puzz *legend)
 {
-	legend.sq_size = ft_sqrt(4 * legend.pnbr);
+	legend->sq_size = ft_sqrt(4 * legend->pnbr);
 	legend = eb_calculator(pieces, legend);
-	ft_memdel((void *)&legend.map);
+	ft_memdel((void *)&legend->map);
 }
